# contracts/openapi/content-notification-service.yaml
openapi: 3.1.0
info:
  title: Content-Notification Service API
  version: 1.6.0
  description: >
    MentalBridge Content/Notification bounded context.
    Manages reviewed self-help resources, reviewed Vietnam safety-directory
    records, in-app notifications, and delivery preferences. The removed legacy
    hotline table is not restored; directory records use the ADR 0018 review,
    provenance, coverage, and 90-day verification model.
servers:
  - url: http://localhost:3003
    description: Local development
...
paths:
...
  /api/v1/resources:
    x-mentalbridge-status: implemented
    get:
      x-mentalbridge-operation-status: implemented
      tags: [Resources]
      summary: List active reviewed published self-help resources
      operationId: listResources
      parameters:
        - name: locale
          in: query
          schema:
            type: string
            example: vi-VN
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/ResourceCategory"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: >
            Published reviewed resources. Returns empty array when none match.
            Returns fallback: unavailable with neutral message when service is unreachable.
            No hotline number or emergency dispatch claim is included.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
              examples:
                published:
                  value:
                    data:
                      - id: 123e4567-e89b-12d3-a456-426614174000
                        category: BREATHING
                        locale: vi-VN
                        title: Kỹ thuật thở 4-7-8
                        summary: Kỹ thuật thở giúp giảm căng thẳng
                        externalUrl: null
                        status: PUBLISHED
                        reviewedAt: "2024-01-15T10:00:00Z"
                        createdAt: "2024-01-15T09:00:00Z"
                        updatedAt: "2024-01-15T10:00:00Z"
                    count: 1
                empty:
                  value:
                    data: []
                    count: 0
                unavailable:
                  value:
                    data: []
                    count: 0
                    fallback: unavailable
                    message: Tài nguyên hỗ trợ tạm thời không khả dụng. Vui lòng thử lại sau.
        "400":
          $ref: "#/components/responses/RequestRejected"

    post:
      x-mentalbridge-operation-status: implemented
      tags: [Resources]
      summary: Create a new resource (Admin)
      operationId: createResource
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceRequest"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceSummary"
              example:
                id: 123e4567-e89b-12d3-a456-426614174000
                category: BREATHING
                locale: vi-VN
                title: Kỹ thuật thở 4-7-8
                summary: Kỹ thuật thở giúp giảm căng thẳng và lo âu
                externalUrl: null
                status: DRAFT
                reviewedAt: null
                createdAt: "2024-01-15T10:00:00Z"
                updatedAt: "2024-01-15T10:00:00Z"
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/resources/admin/list:
    x-mentalbridge-status: implemented
    get:
      tags: [Resources]
      summary: List resources in every lifecycle state (Admin)
      operationId: listAdminResources
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Locale"
        - $ref: "#/components/parameters/Category"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ResourceStatus" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: One page of admin-visible resources
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceListResponse" }
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /api/v1/resources/admin/{id}:
    x-mentalbridge-status: implemented
    get:
      tags: [Resources]
      summary: Get any resource lifecycle state (Admin)
      operationId: getAdminResource
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Full admin resource detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminResourceDetail" }
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/resources/{id}:
    x-mentalbridge-status: implemented
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Resources]
      summary: Get an active reviewed resource for one locale
      operationId: getResource
      parameters:
        - $ref: "#/components/parameters/Locale"
        - name: contentVersion
          in: query
          required: false
          description: >
            Exact non-negative content version requested by a persisted consumer reference.
            When present, a different current version returns 404 rather than substituting content.
          schema:
            type: string
            pattern: "^(0|[1-9][0-9]*)$"
            maxLength: 19
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicResourceDetail"
        "400": { $ref: "#/components/responses/RequestRejected" }
        "404": { $ref: "#/components/responses/NotFound" }

    patch:
      tags: [Resources]
      summary: Update a DRAFT resource (Admin)
      operationId: updateResource
      security:
        - BearerAuth: []
      parameters:
        - name: version
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceRequest"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceSummary"
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [Resources]
      summary: Delete a DRAFT resource (Admin)
      operationId: deleteResource
      security:
        - BearerAuth: []
      parameters:
        - name: version
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/resources/{id}/publish:
    x-mentalbridge-status: implemented
    post:
      tags: [Resources]
      summary: Publish a reviewed DRAFT resource (blocked pending MB-251 policy approval)
      operationId: publishResource
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/ReviewApprovalRequired" }
  /api/v1/resources/{id}/archive:
    x-mentalbridge-status: implemented
    post:
      tags: [Resources]
      summary: Archive a PUBLISHED resource (Admin)
      operationId: archiveResource
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Resource archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceSummary"
        "400": { $ref: "#/components/responses/RequestRejected" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque key reused only for an unchanged logical create request by this actor.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9_-]+$"
    Locale:
      name: locale
      in: query
      schema:
        type: string
        pattern: "^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$"
        default: vi-VN
    Category:
      name: category
      in: query
      schema: { $ref: "#/components/schemas/ResourceCategory" }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, format: uuid }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by identity-service. Resource administration operations
        require the ADMIN role.
  responses:
    Unauthorized:
      description: Authentication is required or the credential is invalid
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    Forbidden:
      description: Authenticated actor is not an administrator
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    ServiceUnavailable:
      description: Resource administration is temporarily unavailable
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    IdempotencyConflict:
      description: The actor reused an idempotency key with a different create payload
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    ReviewApprovalRequired:
      description: Publish is unavailable until an approved MB-251 review decision is recorded
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: https://mentalbridge.io/errors/RESOURCE_NOT_FOUND
            title: Resource not found
            status: 404
            code: RESOURCE_NOT_FOUND
            correlationId: 123e4567-e89b-12d3-a456-426614174000
...
    ValidationError:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: https://mentalbridge.io/errors/VALIDATION_ERROR
            title: Validation failed
            status: 422
            code: VALIDATION_ERROR
            correlationId: 123e4567-e89b-12d3-a456-426614174000
            fieldViolations:
              - field: title
                message: Title is required
              - field: category
                message: Invalid category
    RequestRejected:
      description: Invalid query parameter
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: https://mentalbridge.io/errors/REQUEST_REJECTED
            title: Request rejected
            status: 400
            code: REQUEST_REJECTED
            correlationId: 123e4567-e89b-12d3-a456-426614174000
...
  schemas:
...
    ResourceCategory:
      type: string
      enum: [BREATHING, MEDITATION, ARTICLE, VIDEO, JOURNALING, COMMUNITY]
    ResourceStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
...
    ResourceKind:
      type: string
      enum: [LEARNING, PRACTICE, HABIT, ACTION, REFLECTION]
...
    ResourceSummary:
      type: object
      required:
        - id
        - category
        - resourceKind
        - interactionType
        - repeatability
        - completionMode
        - streakEligible
        - expectedDurationMinutes
        - cooldownDays
        - recommendedFrequencyPerWeek
        - planTags
        - locale
        - title
        - summary
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/ResourceCategory"
        resourceKind:
          $ref: "#/components/schemas/ResourceKind"
        interactionType:
          $ref: "#/components/schemas/ResourceInteractionType"
        repeatability:
          $ref: "#/components/schemas/ResourceRepeatability"
        completionMode:
          $ref: "#/components/schemas/ResourceCompletionMode"
        streakEligible:
          type: boolean
        expectedDurationMinutes:
          type: integer
          minimum: 1
        cooldownDays:
          type: integer
          minimum: 0
        recommendedFrequencyPerWeek:
          type: integer
          minimum: 1
        planTags:
          type: array
          uniqueItems: true
          items:
            type: string
        locale:
          type: string
        title:
          type: string
        summary:
          type: string
        externalUrl:
          type: string
          nullable: true
        sourceOrganization:
          type: string
          nullable: true
          description: Simple accountable source label for catalogue cards.
        status:
          $ref: "#/components/schemas/ResourceStatus"
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PublicResourceDetail:
      allOf:
        - $ref: "#/components/schemas/ResourceSummary"
        - type: object
          required:
            - contentVersion
            - contentBody
            - effectiveAt
            - expiresAt
            - structuredContent
            - interactionConfig
            - safetyNotes
            - sourceRetrievedAt
            - sourceContentHash
            - contentVersionLabel
            - sourceReviewStatus
          properties:
            contentVersion:
              type: string
              pattern: "^(0|[1-9][0-9]*)$"
              maxLength: 19
              description: Exact immutable resource version represented by this detail response.
            contentBody:
              type: string
              nullable: true
            sourceTitle:
              type: string
              nullable: true
            sourceUrl:
              type: string
              format: uri
              pattern: "^https?://"
              nullable: true
            sourceReviewNote:
              type: string
              nullable: true
            structuredContent:
              type: object
              additionalProperties: true
            interactionConfig:
              type: object
              additionalProperties: true
            safetyNotes:
              type: array
              items:
                type: string
            sourceRetrievedAt:
              type: string
              format: date-time
              nullable: true
            sourceContentHash:
              type: string
              pattern: "^[a-f0-9]{64}$"
              nullable: true
            contentVersionLabel:
              type: string
            sourceReviewStatus:
              type: string
              enum: [REVIEWED, REVIEW_REQUIRED, NEEDS_SOURCE_REVIEW]
            effectiveAt:
              type: string
              format: date-time
              nullable: true
            expiresAt:
              type: string
              format: date-time
              nullable: true
    AdminResourceDetail:
      allOf:
        - $ref: "#/components/schemas/PublicResourceDetail"
        - type: object
          required: [reviewedBy, version]
          properties:
            reviewedBy:
              type: string
              format: uuid
              nullable: true
            version:
              type: integer
    ResourceListResponse:
      type: object
      required: [data, count]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ResourceSummary"
        count:
          type: integer
        nextCursor:
          type: string
          format: uuid
          description: Cursor for the next page; absent when no more results.
        fallback:
          type: string
          enum: [unavailable]
          description: >
            Present only when Content/Notification is unavailable.
            No hotline number or emergency dispatch claim is included.
        message:
          type: string
          description: Neutral fallback message when fallback is unavailable.
    CreateResourceRequest:
      type: object
      required: [category, title, summary]
      properties:
        category:
          $ref: "#/components/schemas/ResourceCategory"
        locale:
          type: string
          default: vi-VN
        title:
          type: string
          maxLength: 255
        summary:
          type: string
        contentBody:
          type: string
          nullable: true
        externalUrl:
          type: string
          format: uri
          pattern: "^https?://"
          nullable: true
          description: VIDEO resources require a verified HTTPS YouTube URL.
        sourceOrganization:
          type: string
          maxLength: 200
          nullable: true
        sourceTitle:
          type: string
          maxLength: 500
          nullable: true
        sourceUrl:
          type: string
          format: uri
          pattern: "^https?://"
          nullable: true
        sourceReviewNote:
          type: string
          nullable: true
        effectiveAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
...